Set intel_iommu=on in kernel command line
In Nixpkgs kernels, Intel IOMMUs are disabled by
default (CONFIG_INTEL_IOMMU_DEFAULT_ON=n), so it needs to be
explicitly enabled on the kernel command line.  This was already done
in the rootfs run make target, which is what I mostly use for testing,
which is why I hadn't noticed until now.

We could also set CONFIG_INTEL_IOMMU_DEFAULT_ON, but I think it's
better to avoid changing our kernel configuration too much at the
moment.  (This might change in future if we want to go all in.)

No changes are required for AMD IOMMUs, which are enabled by default.

Reported-by: Vadim Likholetov <vadim.likholetov@unikie.com>
Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-Id: <20220922075629.247058-2-hi@alyssa.is>