host/rootfs: move ext mounting to s6-rc service
Broadly, it makes sense to do things in s6-rc rather than mdev
commands where possible, since mdev commands are not run in parallel.
That's unlikely to make a big difference in this case, but we might
want to do more things before mounting as well (e.g. growing the
filesystem) and I think it makes sense to do that here.

Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-Id: <20221010232909.1953738-12-hi@alyssa.is>