host/rootfs: use a bigger test ext partition
So that we can test things that write to disk.

Unlinking the file prior to starting the VM ensures that it won't
stick around once it's no longer needed.

I disable make's printing of the commands in favour of the shell's,
because the cp can take some time on filesystems that don't support
reflinks, and it's nice to see which command is currently running
rather than all of them at once as make would show.

Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-Id: <20221010232909.1953738-14-hi@alyssa.is>