host/initramfs/extfs.nix: tar2ext4 -> mkfs.ext4 -d
tar2ext4 is designed for making read-only, minimally-sized ext4
images.  That's great for the Spectrum host system's root filesystem,
but it doesn't make sense for the user data partition, because even if
we mark it read/write after running tar2ext4, there will be very few
inodes available, so it won't be very useful for storing data.

To fix this, we use the normal tools for creating an ext4 filesystem.
We create a sparse image of an extremely large size to ensure there's
enough space available for metadata and things for the partition to
grow to fit any disk a user might want to to use Spectrum on, then
shrink the image down to its minimum size.

Doing it this way assumes that the build environment will always
support sparse files.  If this turns out not to be the case, we can
use mkfs.ext4's -E resize= option to set the maximum expected disk
size, and use a smaller size for the initial image size.  I haven't
done that here to avoid having to pick that smaller number.

Now that the filesystem is no longer marked as read-only, we need to
fsck it before every mount, or it won't mount any more after the first
time.

Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-Id: <20221010232909.1953738-15-hi@alyssa.is>