release.nix: run shellcheck on build scripts
This will make the binary cache build (which I expect to become more
CI-like over time) fail if shellcheck doesn't pass on any shell
scripts found in the tree.  This will help ensure we are sticking to
POSIX shell features and following shell scripting best practices.

At present, only scripts with names ending in ".sh" are processed, as
we don't have any others, and discovering extensionless scripts would
be quite a bit more complex.  We can worry about that when the time
comes.

Signed-off-by: Alyssa Ross <alyssa.ross@unikie.com>
Message-Id: <20221110112219.459660-2-alyssa.ross@unikie.com>
Reviewed-by: Henri Rosten <henri.rosten@unikie.com>