Documentation: rename "UUID" to "UUID Reference"
The purpose of this document is specifically to provide a reference
table so that developers working with Spectrum can easily look up a
given UUID to find out what it means.

I think "UUID Reference" makes it clearer what the reader will find in
this document, as I'd expect a document just called "UUID" to be
something that explained e.g. what a UUID is.  (I don't think we need
to explain that at all in the Spectrum documentation, as we link to
Wikipedia, which explains it well, and it's background knowledge that
most developers looking to work with Spectrum will likely already
have.)

Signed-off-by: Alyssa Ross <alyssa.ross@unikie.com>
Cc: Jenni Nikolaenko <evgeniia.nikolaenko@unikie.com>
Message-Id: <20221116105423.610445-1-alyssa.ross@unikie.com>