Documentation/scripts/build.sh: avoid unquoted var
Recent versions of shellcheck warn about word splitting on
JEKYLLFLAGS.  To avoid this, let's switch to using "$@" to store
the Jekyll command line we're building up.

This also gives us a consistent way to pass extra arguments to Jekyll.
They just need to be specified on the command line — we can remove
special handling for the destination by just having the caller pass
-d like they would when dealing with Jekyll directly.

Signed-off-by: Alyssa Ross <hi@alyssa.is>