release: don't pass -snapshot to QEMU
This is a remnant from before we made our own temporary file to act as
the installation disk.  Now, it just means that qemu will pointlessly
copy the empty image to TMPDIR.  Then, if that filesystem fills
up (which is not unlikely because there's a good chance it's tmpfs)
and writes start failing, QEMU will pause the VM.  This made it
difficult to test the installer.

Signed-off-by: Alyssa Ross <hi@alyssa.is>