nix/eval-config.nix: fix custom configuration
When I introduced the global src attribute, I broke custom
configurations, since they wouldn't be passing that key in, and nor
should they.  To allow for non-customisable globals like src, we need
to separate them from config.  Here, I've modified eval-config.nix to
take a callback, so that it can provide multiple attributes, and
handle command line arguments itself so that doesn't need to be in
every entry point any more.  This gives us an interface similar to the
NixOS module system's, where a variety of globals are available that
can be pulled out of the passed attribute set as required, but with
the additional advantage that files are usable directly from
nix-build, and support command line arguments.

Signed-off-by: Alyssa Ross <hi@alyssa.is>
Fixes: 0149885 ("nix: centralise source cleaning")