nix: inline source finding in eval-config.nix
Originally, I was thinking every file that needed to operate on the
Spectrum sources would need to import src.nix, but I ended up making
it a global instead.  I think that's nicer to use, but also avoids
re-evaluating the sources for every component in a full build, which I
don't think Nix would memoize due to it being wrapped in a function to
get access to lib (but I didn't test).

Since src.nix therefore shouldn't be accessed anywhere else, it's best
to inline its contents into eval-config.nix, so that it _can't_ be
accessed anywhere else.

Signed-off-by: Alyssa Ross <hi@alyssa.is>