release/checks/wayland: init
This is a smoke test for crosvm vhost-user-gpu.  It starts Weston and
the img/app VM in crosvm, and checks that the hello-wayland window
from the VM appears, using a small Weston plugin.

Hopefully this will make it easier to keep up with upstream crosvm,
because testing crosvm versions can be automated.

Signed-off-by: Alyssa Ross <hi@alyssa.is>