Build make targets directly into destinations
Our image Makefiles are either used for interactive development in a
nix-shell, or as part of a Nix build of Spectrum.  In neither of these
cases is the two step "make; make install" process useful like it is
for packages, because one image from Spectrum somewhere on the
development system isn't something it ever makes sense to do.

By removing the separation between building and installing, and
creating the final output directly in its destination, we can save
some copies of very large files, which should save seconds in each
build, even for the smaller components.

Signed-off-by: Alyssa Ross <hi@alyssa.is>