lib/config.default.nix: pin Nixpkgs
This makes the common case of building Spectrum, using the default
Nixpkgs, much easier.  It's easy to make mistakes setting up NIX_PATH,
so we should leave custom Nixpkgs to the config file or nix command
line.

This doesn't preclude using Flakes to pin Nixpkgs in future, if we end
up going that route.

Signed-off-by: Alyssa Ross <hi@alyssa.is>