host/rootfs: allow growing ext partition to fail
When using "Try Spectrum" from the combined image, the ext partition
will be read only, and resize2fs will fail.  In this situation, we
want to mount it read only, so it's still possible to try out the
built in VMs.  mount(1) will automatically fall back to a read-only
mount if a read-write mount fails, but we need to get far enough to
let it try, so accept that resize2fs might fail and carry on to mount
regardless.

Fixes: 8c350b7 ("host/rootfs: automatically grow user partition")
Signed-off-by: Alyssa Ross <hi@alyssa.is>