release/checks/try.nix: init
This is a regression test for c7f87f3 ("host/rootfs: allow growing ext
partition to fail").  It's the first time we're actually doing
automated tests of a Spectrum boot.  For now, I'm using the NixOS test
framework, but because we're not using NixOS and not setting any NixOS
options, it feels to me like it doesn't actually buy us very much, so
if it doesn't start adding more value as we add more (or more complex)
tests, it might be simpler to just use a shell/execline script for
tests.

Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-Id: <20230526210757.397735-1-hi@alyssa.is>
Reviewed-by: Ryan Lahfa <ryan@lahfa.xyz>