host/initramfs: panic if roothash is missing
We're not going to be able to boot if we don't know what roothash
we're looking for, so abort the boot at that point rather than hanging
forever.

This should never happen on an actual system, but it makes things
clearer in testing if something has gone wrong with the kernel command
line.

Signed-off-by: Alyssa Ross <hi@alyssa.is>