Documentation: fix accidental anchor links
The xref macro is for cross references to internal links within the
documentation, not external links.  xref:https://… results in an
anchor link.

The existing doc-links test didn't catch this, because it operates at
the HTTP level.  But the problem is easy to spot in the sources, so
I've added a new check that just greps for this pattern.

Reported-by: Ryan Lahfa <ryan@lahfa.xyz>
Fixes: 97cceff ("Documentation: add contributing information")
Signed-off-by: Alyssa Ross <hi@alyssa.is>
Message-Id: <20230527151152.2016164-1-hi@alyssa.is>