img/app: shut down when the application exits
As we move towards graphical applications, the intuitive way for
things to work is to have application VMs terminate when the
application exits.  Once the terminal-based applications are gone (or
at least not the main focus), we can repurpose the console device to
be an interactive shell, so debugging is still possible.

This requires actually implementing an orderly shutdown in the VM.
Prior to now, there's just been poweroff -f.

Signed-off-by: Alyssa Ross <hi@alyssa.is>