vm: remove mesa drivers
This includes a revert of commit 1a754374028ced6eea2dd77a2f829b511d2e90ba.

vm-lib/make-vm.nix was not the right place for the Mesa drivers to be
added — it meant that every VM had its own copy of the drivers,
whether it used them or not.  And right now, nothing uses the Mesa
drivers.  This will only be the case if we do GPU passthrough at some
point.  When that happens, we can add them in a more sensible way, in
img/app.

Signed-off-by: Alyssa Ross <hi@alyssa.is>