img/app: run wayland-proxy-virtwl as a service
This reduces duplication between VM definitions, which is important,
because eventually we don't want VMs to need to be defined up front at
all — it should be possible to run applications on demand without
needing to write a VM definition for them.

The copyright text in foot.nix has been adjusted because nothing owned
by Unikie is left in the file.  What's left is probably not
copyrightable at all, but if any of it is, it's mine.

Signed-off-by: Alyssa Ross <hi@alyssa.is>