host/start-vm: find config from symlink in service
This decouples start-vm from where VM configs are stored, and
decouples VM configs from VM names, which will make it possible to
have VMs that are built in to the system, and multiple VMs using the
same config with different environments (e.g. the multiple instances
of the same application).

Signed-off-by: Alyssa Ross <hi@alyssa.is>