host/start-vm: enable -Werror in Nix builds
This would have stopped me committing the warning fixed in
2441673 ("host/start-vm: remove unused import").

TUNGETIFF is cast to unsigned, which removes a warning caused by
musl's idea of what an ioctl request is differing from Linux's[1].

C standard is now c2x, to disable a -Wpedantic warning about square
bracket attributes not being valid in C11.

-Wno-error=attributes is set, because the standard says that unknown
attributes are supposed to be ignored by the compiler.  Having a
single warning about it is fine, but not an error.

[1]: https://inbox.vuxu.org/musl/20200120171425.GL30412@brightrain.aerifal.cx/

Signed-off-by: Alyssa Ross <hi@alyssa.is>