host/start-vm: find config from /run/vm
Having a symlink in the service directory was a nice idea, but since
it has to be different for every service, it's not compatible with
templated services, and would prevent us switching to them.  The only
thing that varies between instances of templated services is the name,
so we have to introduce a mapping of VM name to config directory.

Signed-off-by: Alyssa Ross <hi@alyssa.is>