lib: adjust interface for Nix files
This moves every directly-buildable Nix file in Spectrum from the old
eval-config.nix interface to a new callPackage-based interface.  This
moves us in the direction of stopping files from directly importing
each other, in favor of having a global package set that contains the
Spectrum-specific packages.

Aside from consistency with packages from Nixpkgs, the main advantage
to this is that the packages will be correctly spliced, and so
e.g. lseek being in nativeBuildInputs will now do the right thing.

This is implemented using a scope, so the Spectrum packages are
invisible to Nixpkgs, and are applied on top afterwards, in contrast
to if an overlay was used.

Having a customised package set also paves the way for bringing
modifications to upstream packages (e.g. the Cloud Hypervisor
virtio-gpu patches) into the Spectrum repository, allowing us to use
an unmodified upstream Nixpkgs.

Signed-off-by: Alyssa Ross <hi@alyssa.is>