pkgs: move cloud-hypervisor patch into Spectrum
Previously, this lived in our custom Nixpkgs tree.  Since it would be
nice to move away from having one of those, we need to bring this into
Spectrum directly.

This goes in an overlay because if some package in Nixpkgs uses
Cloud Hypervisor, there's no point ending up with two Cloud Hypervisor
builds in our build closure, one patched and one not.

Signed-off-by: Alyssa Ross <hi@alyssa.is>