vm: send cloud-hypervisor serial output to file
This didn't originally go to a file, because I still had to use the
serial console interactively for debugging sometimes.  Now that things
have settled a bit and the serial console is write-only, having serial
go to a pty is just needlessly inconvenient.

I've updated the crosvm target to put its logs in the same place, so
the path to serial logs only has to be learned once.

Signed-off-by: Alyssa Ross <hi@alyssa.is>