release/checks/wayland: fix false positives
Sometimes the application running inside the crosvm VM can commit
between wc exiting and the VM being stopped.  If this happens,
there'll be an extra byte in the fifo, which will interfere with the
second test.  To fix this, recreate the fifo to ensure it's empty at
the start of the Cloud Hypervisor test.

Fixes: 5a5f892 ("release/checks/wayland: also test cloud-hypervisor")
Signed-off-by: Alyssa Ross <hi@alyssa.is>