lib/nixpkgs.default.nix: update
This is a manual update to the last commit in Nixpkgs that doesn't
have Meson 1.3.0, which introduces a bug that breaks the Spectrum
build[1], for which upstream have not yet accepted a patch.

We can patch it ourselves if it's still not fixed by the time we need
another update, but for now, we don't need anything more recent than
this, so I'll give it a bit more time to get fixed.

[1]: https://github.com/mesonbuild/meson/issues/12585

Signed-off-by: Alyssa Ross <hi@alyssa.is>