Documentation: remove references to mailman-web
Bots have been nonconsensually creating mailman-web accounts for
people to flood them with spam.  Aside from using reCAPTCHA (ugh) or
modifying mailman-web, the only thing I can really do to prevent this
is to disable unauthenticated web subscriptions, and the creation of
mailman-web accounts, which I have now done.  Ideally, it would be
possible to configure it so that creating mailman-web accounts is only
allowed for email addresses that have already subscribed to the list
by sending an email, but that's not currently the case.

This makes Postorius useless to people who don't already have an
account.  Hyperkitty is still useful, but people without existing
accounts won't be able to use it to post to the lists, and the UI
still shows Sign In and Sign Up links, which is pretty confusing when
signups are disabled.  It doesn't really offer anything that
public-inbox doesn't any more.

As a result of these changes, I'm removing all references to
mailman-web from the documentation.

Signed-off-by: Alyssa Ross <hi@alyssa.is>