Documentation: improve prose of communication page
There was a lot of fluff here that got in the way of actually finding
the information people are coming to this page to find.

"You are strongly encouraged to mirror them." is probably confusing to
people who don't know what mirroring this, so rephrase this to make it
clearer that this isn't something people need to worry about in that
case.

Signed-off-by: Alyssa Ross <hi@alyssa.is>