host/start-vm: add run-spectrum-vm to tests
release/checks/start-vm.nix wasn't actually imported anywhere, so it
didn't do anything unless it was built on its own.

We could have listed it in release/checks/default.nix, but I think it
makes more sense to have it in start-vm's passthru.tests, so it's more
obvious that it should be run when changing start-vm.

Fixes: 009a036 ("release/checks/start-vm.nix: init")
Signed-off-by: Alyssa Ross <hi@alyssa.is>