host/start-vmm: use a Meson option for panic
This is a much more meson-y way to handle this, and it should make it
easier to use Meson's new Cargo wraps.

I don't like that it's the non-native dependencies that need to have
this option, but it seems to fit into Meson much better this way, so
it is what it is.

Signed-off-by: Alyssa Ross <hi@alyssa.is>