Access PACKAGES through the shell instead of make
The command line here was still getting too long for the shell.
Previously, I assumed this was make-erofs.sh's command line, but it
was probably actually the command line for the shell invoked by make
with sh -c.  To reduce the length of this, don't have make pass the
whole PACKAGES list to the shell.  Instead have the shell read it from
the environment itself.

PACKAGES is not defined in the Makefile, so it has to be provided
either in the environment or on the make command line.  POSIX
specifies that macros defined on the make command line are added to
the environment, so this should still work in both cases.

Signed-off-by: Alyssa Ross <hi@alyssa.is>