host/rootfs: add a timeout for waiting for kvm
/dev/kvm appears at some point after boot when the appropriate module
is loaded.  This means we need a way to distinguish between the module
not having loaded yet and kvm being completely unsupported, because
otherwise we'll be waiting indefinitely (locking s6-rc forever).

Spectrum running in QEMU TCG on the slowest computer I have access to
takes a bit under 20 seconds, so 40 should be a reasonable amount of
time to give up after.

Signed-off-by: Alyssa Ross <hi@alyssa.is>