host/rootfs: always share a single dir with VMs
This makes everything simpler.  Having multiple virtio-fs devices
would be inefficient, and would raise the question of how VMs should
identify which one is which, and it doesn't seem to buy us anything
over bind mounts into a single directory.

As part of this change, I've removed the static configuration of
shared directories rather than adapting it.  It would be possible to
implement it, but I think that (like much of the rest of Spectrum)
we'll be able to better design static configuration if we get dynamic
stuff working first.

Currently, VMs get this single shared directory whether they want it
or not.  We could make it optional in future.

mount now comes from util-linux, because if you do mount --bind -o ro
with busybox mount, it'll just ignore the -o ro part.  Terrifying.

Signed-off-by: Alyssa Ross <hi@alyssa.is>