host/start-vmm: provide Wayland unconditionally
There's always going to be a need for getting a shell a VM to see
what's going on.  Using a terminal emulator running on the host for
this is probably not a good idea, because terminal emulators are
unlikely to be robust against a malicious application doing weird
things with escape sequences, so the best way to do this is probably
going to be to have every VM have a Wayland terminal emulator in it,
and some way to tell the VM to start it.  This means that every VM
will need a cross-domain virtio-gpu device.

This will also make it easier to include Wayland support in shared VM
images, because there's no need to handle the case where there's no
device.

Signed-off-by: Alyssa Ross <hi@alyssa.is>