img/app: install libs, bins needed for AppImages
Unfortunately, AppImages are not hermetic, and require a poorly
defined set of binaries and libraries to be provided by the system.
We can reuse Nixpkgs's list for this.  Hopefully we don't need
anything from packages in etc, apart from the the SSL certs we already
have.  I've switched those to be installed in the normal etc location,
so that we don't have to try to insert anything of our own into usr.

D-Bus and lots of the busybox commands are provided by the AppImage
environment, so we can remove the copies of those we previously
installed.  (We should use the ones from the AppImage environment in
case AppImages rely on GNUisms etc.)

Now that we're using GNU awk instead of busybox awk, we unfortunately
need to pass it the non-standard -P (POSIX) flag to get the standard
behaviour of interpreting "0x"-prefixed strings as hexadecimal
numbers.

Signed-off-by: Alyssa Ross <hi@alyssa.is>