Pass --no-reread --no-tell-kernel to sfdisk
We're generating image files, so there's no need to tell the kernel to
reread the partition table, and doing so will cause sfdisk to call
sync(), which could waste a lot of time depending on what else is
happening on the system.

I don't think --no-reread matters as much, but we don't need it
either.

Signed-off-by: Alyssa Ross <hi@alyssa.is>