host/rootfs: prefix VMs with source
The user data partition should only be for user VMs, which means we'll
need a concept of system VMs.  That means we have to figure out what
to do about name collisions.  The easiest thing to do is to just
prefix the VM name with its source.  Right now, the only prefix is
"user", but now it will be easy to add other sources.  In future, we
could maybe also replace "user" with a partition UUID or something, to
allow for multiple sources of VMs.

Signed-off-by: Alyssa Ross <hi@alyssa.is>