host/rootfs: move API and vsock sockets to /run/vm
Having them in the service directory isn't going to work if we start
having transient VMs that aren't services.

start-vmm now always uses absolute paths, because its CWD is no longer
the directory containing the sockets.  The VM name validation is moved
into vm_config(), as it's now the only place that does anything with
the VM name itself, as opposed to the whole directory.

Signed-off-by: Alyssa Ross <hi@alyssa.is>