host/rootfs: start VMM dependencies in parallel
Now that the VSOCK unix socket is in /run/vm, the
xdg-desktop-portal-spectrum-host service no longer needs to wait for
the VMM service directory to be created.  Because of this, and the
fact that none of these services have dependencies on each other (at
least until the VMM is started), we can create them in parallel.

Signed-off-by: Alyssa Ross <hi@alyssa.is>