host/start-vmm: fix cleaning up when VMM exits
Cloud Hypervisor only assigns IDs to devices once the VM has booted.
Given VMs are all set up at the same time, this is very unlikely to be
the case.  When the VM hadn't booted, its ID was being recorded as the
empty string (because jq hadn't been given -e, so wouldn't fail when
there was no response from Cloud Hypervisor), which would then mean
that the router net device wouldn't be removed when the client VMM
exited.

To fix this, assign IDs based on the tap name, rather than expecting
Cloud Hypervisor to do it.  This way, it doesn't matter that the
router VM isn't booted when the client VM is set up.

Fixes: a2d362c ("host: allow VMs to be powered off")
Signed-off-by: Alyssa Ross <hi@alyssa.is>