vm/make-vm.nix: use rootfs's app image
rootfs's version is (currently) not the same as the one you get from
img/app, because it passes the terminfo from its own foot build.  This
meant that basePaths was not actually the set of paths that existed in
the image, so it was possible to end up with paths that were neither
in the base image, nor in the application-specific image.

Fixes: 2b00830 ("img/app: extract from appvm-{lynx,catgirl}")
Signed-off-by: Alyssa Ross <hi@alyssa.is>