crosvm: Rename `--vhost-user-gpu` flag
crosvm was producing warnings when using `--vhost-user-gpu` and
`--vhost-user-fs`, of which the former is fixable, but the latter
requires some more work.

To mitigate the former, I have adjusted the `crosvm` invocations to look
something like `--vhost-user $DEVICE,socket=$PATH`, where `$DEVICE` is,
in this case, `gpu`, and `$PATH` is the path to the Unix socket.

Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>