crosvm: Rename `--vhost-user-gpu` flag
crosvm was producing warnings when using `--vhost-user-gpu` and
`--vhost-user-fs`, of which the former is fixable, but the latter
requires some more work. The latter will be patched later on.

To mitigate the former, I have adjusted the `crosvm` invocations to look
something like `--vhost-user $DEVICE,socket=$PATH`, where `$DEVICE` is,
in this case, `gpu`, and `$PATH` is the path to the Unix socket.

Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>
Message-ID: <20241108181450.1424488-1-shymega@shymega.org.uk>
Reviewed-by: Alyssa Ross <hi@alyssa.is>
Tested-by: Alyssa Ross <hi@alyssa.is>
Signed-off-by: Alyssa Ross <hi@alyssa.is>