release/live: remove console override
A serial console shouldn't be required in normal use.  It's nicer not
to interfere with the kernel's default behavior, especially on other
architectures.

We do still need a console for release/checks/try.nix to work.  For
now it stays as hardcoded to ttyS0, because this test only works on
x86_64 at the moment anyway.

Signed-off-by: Alyssa Ross <hi@alyssa.is>