scripts/run-qemu.sh: use TCG on aarch64 for NV
KVM does not support nested virtualization on aarch64 yet, so it's
either TCG, or not being able to actually test things properly.

Callers of run-qemu.sh indicate whether they need nested
virtualization using -machine virtualization=on.  On aarch64, that's
forwarded to QEMU, and KVM is disabled.  On other platforms, the
virtualization parameter is stripped, as it's aarch64-specific.

Signed-off-by: Alyssa Ross <hi@alyssa.is>