host/rootfs: disable s6-notifyoncheck timeout
The default timeout is 7010ms, which is likely to be hit running on
TCG (which we have to use for development on aarch64).  I don't think
there's really any reason we'd ever want this to time out, and if it
does, it means a VM will just never become available.

Signed-off-by: Alyssa Ross <hi@alyssa.is>