release/live: boot via UKI
This will allow us to use the new .dtbauto mechanism to load device
trees on aarch64.  It will also make systemd-boot dual boot
installation easier, as just a single file needs to be copied to the
ESP.

This means that we're back to chainloading for "Try Spectrum" in the
installer.  That was previously switched away from in
b1ff26e ("img/combined: avoid chainloading in GRUB") because of
firmware issues.  There's not really an alternative if we want to use
UKIs though, so hopefully things will have got better since then?

Signed-off-by: Alyssa Ross <hi@alyssa.is>